Reorder internal modules (#1117)
* Dont export internal `validateSampleConfig` from sample

* Move `createLinkNode` to `createNode` module

* Move is to `isMethod`

* Create meaningful `validate` module

* Move `assert` and `deprecate` to validate module

* Move `is.*` methods to `is` module

Essentially a flip with validate module

* Move watchUnit to createWatch module

* Merge subscription module to createWatch

* Merge observableSymbol to fromObservable module

* Merge own to createNode module

* Fix import

* Fix more imports
