+++ src/effector/combine.ts
@@ -4,10 +4,9 @@ import {createStateRef, addRefOp} from './stateRef'
 import {mov, calc, read, userFnCall} from './step'
 import {processArgsToConfig} from './config'
 import {getStoreState, setMeta} from './getter'
-import {is, isFunction, isObject, isVoid} from './is'
+import {is, isFunction, isObject, isVoid, assert} from './validate'
 import {generateErrorTitle, unitObjectName} from './naming'
-import {createLinkNode} from './forward'
-import {assert} from './throw'
+import {createLinkNode} from './createNode'
 import {readTemplate} from './region'
 import {forIn} from './collection'
 import {MAP, REG_A, VALUE} from './tag'
