+++ src/effector/createEffect.ts
@@ -2,13 +2,11 @@ import type {Unit, Stack} from './index.h'
 import type {Effect} from './unit.h'
 import {calc, run} from './step'
 import {getForkPage, getGraph, getMeta, getParent, setMeta} from './getter'
-import {own} from './own'
-import {createNode} from './createNode'
+import {createNode, own} from './createNode'
 import {launch, setForkPage, forkPage, isWatch, type EffectorQueue} from './kernel'
 import {createStore, createEvent} from './createUnit'
 import {createDefer} from './defer'
-import {isObject, isFunction} from './is'
-import {assert} from './throw'
+import {isObject, isFunction, assert} from './validate'
 import {EFFECT} from './tag'
 import {add} from './collection'
 import {flattenConfig} from './config'
