+++ src/effector/createWatch.ts
@@ -1,10 +1,16 @@
 import {clearNode} from './clearNode'
 import {createNode} from './createNode'
-import type {Subscription, Unit, Compute, MovStoreToRegister} from './index.h'
-import {step} from './step'
-import {Scope} from './unit.h'
-import {addUnsubscribe} from './subscription'
-import {is} from './is'
+import {callStack} from './caller'
+import type {
+  Subscription,
+  Unit,
+  Compute,
+  MovStoreToRegister,
+  NodeUnit,
+} from './index.h'
+import {run, compute, mov} from './step'
+import type {Scope} from './unit.h'
+import {is, isFunction, assert} from './validate'
 
 export function createWatch<T>({
   unit,
@@ -17,15 +23,13 @@ export function createWatch<T>({
   scope?: Scope
   batch?: boolean
 }): Subscription {
-  const seq: (Compute | MovStoreToRegister)[] = [
-    step.run({fn: value => fn(value)}),
-  ]
+  const seq: (Compute | MovStoreToRegister)[] = [run({fn: value => fn(value)})]
   if (batch) {
-    seq.unshift(step.compute({priority: 'sampler', batch: true}))
+    seq.unshift(compute({priority: 'sampler', batch: true}))
   }
   if (is.store(unit)) {
     seq.unshift(
-      step.mov({
+      mov({
         store: (unit as any).stateRef,
         to: 'stack',
       }),
@@ -59,21 +63,20 @@ export function createWatch<T>({
       unsubs.forEach(u => u())
     })
   } else {
-    const node = createNode({
-      node: seq,
-      parent: units,
-      family: {owners: units},
-    })
-    return addUnsubscribe(() => {
-      clearNode(node)
-    })
+    return createSubscription(
+      createNode({
+        node: seq,
+        parent: units,
+        family: {owners: units},
+      }),
+    )
   }
 }
 
 function prepareSeq(seq: (Compute | MovStoreToRegister)[], unit: any) {
   if (is.store(unit)) {
     return [
-      step.mov({
+      mov({
         store: (unit as any).stateRef,
         to: 'stack',
       }),
@@ -82,4 +85,31 @@ function prepareSeq(seq: (Compute | MovStoreToRegister)[], unit: any) {
   }
 
   return seq
+}
+
+export const watchUnit = (
+  unit: NodeUnit,
+  handler: (payload: any) => any,
+): Subscription => {
+  assert(isFunction(handler), '.watch argument should be a function')
+  return createSubscription(
+    createNode({
+      scope: {fn: handler},
+      node: [run({fn: callStack})],
+      parent: unit,
+      meta: {op: 'watch'},
+      family: {owners: unit},
+      regional: true,
+    }),
+  )
+}
+
+export const createSubscription = (node: NodeUnit): Subscription =>
+  addUnsubscribe(() => clearNode(node))
+
+const addUnsubscribe = (callback: () => void): Subscription => {
+  const subscription: Subscription = () => callback()
+  subscription.unsubscribe = () => callback()
+
+  return subscription
 }
