+++ src/effector/index.ts
@@ -13,7 +13,7 @@ export {createEffect} from './createEffect'
 export {createNode} from './createNode'
 export {createStore, createEvent} from './createUnit'
 export {fromObservable} from './fromObservable'
-export {is} from './is'
+export {is} from './validate'
 export {launch, setInspector} from './kernel'
 export {merge} from './merge'
 export {restore} from './restore'
