--- src/effector/observable.ts
@@ -1,2 +1,0 @@
-export const observableSymbol =
-  (typeof Symbol !== 'undefined' && Symbol.observable) || '@@observable'
