+++ src/effector/restore.ts
@@ -1,9 +1,8 @@
 import {createStore} from './createUnit'
-import {is} from './is'
+import {is, assert} from './validate'
 import {forIn} from './collection'
 import {getParent} from './getter'
-import {createLinkNode} from './forward'
-import {assert} from './throw'
+import {createLinkNode} from './createNode'
 import {generateErrorTitle} from './naming'
 
 export function restore(obj: any, defaultState: any, config?: any) {
