+++ src/effector/split.ts
@@ -1,15 +1,13 @@
 import type {DataCarrier} from './unit.h'
 import type {Cmd, Stack} from './index.h'
-import {is, isFunction, isObject, assertTarget} from './is'
+import {is, isFunction, isObject, assertTarget, assert} from './validate'
 import {add, forIn, includes} from './collection'
 import {addRefOp, createStateRef} from './stateRef'
-import {createLinkNode} from './forward'
 import {processArgsToConfig} from './config'
 import {compute, userFnCall, calc, read} from './step'
-import {createNode} from './createNode'
+import {createNode, createLinkNode} from './createNode'
 import {launch, type EffectorQueue} from './kernel'
 import {getStoreState} from './getter'
-import {assert} from './throw'
 import {createEvent} from './createUnit'
 import {applyTemplate} from './template'
 import {createSampling} from './sample'
